+++ CHANGELOG.md
@@ -4,6 +4,7 @@ feat(refactor interface):
 - remove storeName, defaultValue from `init` signature and add it to store
 - resetable store and `reset` method for notifier
 - HiveJsonStore optional `toJson`
+- PersistentStore provide `defaultValue` getter
 
 ## [1.0.0] - 2022-08-04